Move Dir#ReduxPractice
