+++ src/components/ReduxPractice/App.js
@@ -1,0 +1,7 @@
+import React, { Component } from 'react';
+
+export default class App extends Component {
+  render() {
+    return <h2>hello</h2>;
+  }
+}
